Moved UpdateStatus enum and UpdateInfo to separate files
